feat/recipe-modal-implementation
* refactored the structure of the project
