+++ App.tsx
@@ -4,11 +4,11 @@ import {NavigationContainer} from '@react-navigation/native';
 import 'react-native-gesture-handler';
 import {Provider} from 'react-redux';
 import {store} from './src/app/store/store';
-import MenuScreen from './src/app/screens/menu-screen/menu';
-import IntroScreen from './src/app/screens/intro-screen/intro-screen';
-import RecipeListScreen from './src/app/screens/recipe-list-screen/recipe-list';
-import MyRecipesScreen from './src/app/screens/my-recipes-screen/my-recipes-screen';
-import LearnBasicsScreen from './src/app/screens/learn-basics-screen/learn-basics-screen';
+import MenuScreen from './src/app/screens/MenuScreen';
+import IntroScreen from './src/app/screens/IntroScreen';
+import RecipeListScreen from './src/app/screens/AllRecipesScreen';
+import MyRecipesScreen from './src/app/screens/UserRecipesScreen';
+import LearnBasicsScreen from './src/app/screens/BasicsScreen';
 import {ScreensEnum} from './src/app/enums/screens.enum';
 
 const Stack = createStackNavigator();
