--- src/app/components/custom-button/custom-button.interface.ts
@@ -1,5 +1,0 @@
-export interface ButtonPropsInterface {
-    text: string;
-    onPress: () => void;
-    style?: HTMLStyleElement;
-}
