--- src/app/components/custom-modal/custom-modal-recipe-content/custom-modal-recipe-content.styles.ts
@@ -1,22 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-export const styles = StyleSheet.create({
-    container: {
-        width: '100%',
-        height: 'auto',
-        paddingLeft: 6,
-        paddingRight: 6
-    },
-    sectionWrapper: {
-        marginTop: 20,
-        marginBottom: 20
-    },
-    ingredientsHeader: {
-        marginBottom: 20
-    },
-    ingredientWrapper: {
-        display: 'flex',
-        flexDirection: 'row',
-        justifyContent: 'space-between'
-    }
-});
