--- src/app/components/custom-modal/custom-modal.interface.ts
@@ -1,5 +1,0 @@
-export interface CustomModalReduxStateInterface {
-    modal: {
-        isModalVisible: boolean;
-    };
-}
