--- src/app/components/custom-text/custom-text.styles.ts
@@ -1,7 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-export const styles = StyleSheet.create({
-    textStyles: {
-        textAlign: 'center'
-    }
-});
