--- src/app/components/fullscreen-container/fullscreen-container.styles.ts
@@ -1,9 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-export const styles = StyleSheet.create({
-    container: {
-        width: '100%',
-        height: '100%',
-        display: 'flex'
-    }
-});
