+++ src/app/screens/BasicsScreen.tsx
@@ -1,0 +1,9 @@
+import React, {FunctionComponent} from 'react';
+import {Text, View} from 'react-native';
+
+const BasicsScreen: FunctionComponent = (): React.ReactElement =>
+    <View>
+        <Text>Learn screen works!</Text>
+    </View>;
+
+export default BasicsScreen;
