+++ src/app/screens/UserRecipesScreen.tsx
@@ -1,0 +1,9 @@
+import React, {FunctionComponent} from 'react';
+import {Text, View} from 'react-native';
+
+const UserRecipesScreen: FunctionComponent = (): React.ReactElement =>
+    <View>
+        <Text>My recipes screen works!</Text>
+    </View>;
+
+export default UserRecipesScreen;
