--- src/app/screens/intro-screen/intro-screen.interface.ts
@@ -1,3 +1,0 @@
-export interface IntroScreenPropsInterface {
-    navigation: any;
-}
