--- src/app/screens/learn-basics-screen/learn-basics-screen.tsx
@@ -1,15 +1,0 @@
-import React, {Component, ReactElement} from 'react';
-import {Text, View} from 'react-native';
-// ToDo interfaces
-
-class LearnBasicsScreen extends Component {
-    public render(): ReactElement {
-        return (
-            <View>
-                <Text>Learn screen works!</Text>
-            </View>
-        );
-    }
-}
-
-export default LearnBasicsScreen;
