--- src/app/screens/menu-screen/menu-item/menu-item.interface.ts
@@ -1,9 +1,0 @@
-import {ScreensEnum} from '../../../enums/screens.enum';
-import {Image} from 'react-native';
-
-export interface MenuItemPropsInterface {
-    navigation: any;
-    screenToNavigate: ScreensEnum;
-    title: string;
-    image: Image;
-}
