--- src/app/screens/menu-screen/menu.interface.ts
@@ -1,12 +1,0 @@
-import {ScreensEnum} from '../../enums/screens.enum';
-
-export interface MenuItemListInterface {
-    id: string;
-    title: string;
-    image: HTMLElement;
-    screenToNavigate: ScreensEnum;
-}
-
-export interface MenuPropsInterface {
-    navigation: any;
-}
