--- src/app/screens/menu-screen/menu.styles.ts
@@ -1,10 +1,0 @@
-import {StyleSheet} from 'react-native';
-import {ColorsEnum} from '../../enums/colors.enum';
-
-export const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        justifyContent: 'center',
-        backgroundColor: ColorsEnum.LIGHT_GREEN
-    }
-});
