--- src/app/screens/my-recipes-screen/my-recipes-screen.tsx
@@ -1,14 +1,0 @@
-import React, {Component} from 'react';
-import {Text, View} from 'react-native';
-// ToDo interfaces
-class MyRecipesScreen extends Component {
-    public render() {
-        return (
-            <View>
-                <Text>My recipes screen works!</Text>
-            </View>
-        );
-    }
-}
-
-export default MyRecipesScreen;
