--- src/app/screens/recipe-list-screen/recipe-list-item/recipe-list-item.interface.ts
@@ -1,8 +1,0 @@
-import {RecipeListItemInterface} from '../../../interfaces/recipe-item.interface';
-
-export interface RecipeListItemPropsInterface {
-    onPress: () => {};
-    item: {
-        item: RecipeListItemInterface;
-    };
-}
