--- src/app/screens/recipe-list-screen/recipe-list-navbar/recipe-list-navbar.interface.ts
@@ -1,3 +1,0 @@
-export interface RecipeListNavbarPropsInterface {
-    navigation: any;
-}
