--- src/app/screens/recipe-list-screen/recipe-list.interface.ts
@@ -1,19 +1,0 @@
-import {Dispatch} from 'redux';
-import {RecipeListItemInterface} from '../../interfaces/recipe-item.interface';
-
-export interface RecipeListPropsInterface {
-    modal: boolean;
-    dispatch: Dispatch;
-    navigation: any;
-    //later
-    recipeList: {
-        allRecipes: RecipeListItemInterface[];
-    };
-}
-
-export interface RecipeListReduxStateInterface {
-    modal: {
-        isModalVisible: boolean;
-    };
-    recipe: RecipeListItemInterface;
-}
